+++ src/main/java/roomescape/member/controller/AuthController.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.global.auth.AuthenticationPrincipal;
 import roomescape.global.auth.CookieUtil;
+import roomescape.global.auth.LoginMember;
 import roomescape.member.dto.request.AuthRequest.LoginRequest;
-import roomescape.member.dto.request.LoginMember;
 import roomescape.member.dto.response.AuthResponse.LoginResponse;
 import roomescape.member.service.AuthService;
