+++ src/main/java/roomescape/member/service/AuthService.java
@@ -3,9 +3,9 @@ package roomescape.member.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import roomescape.global.auth.JwtUtil;
+import roomescape.global.auth.LoginMember;
 import roomescape.global.error.exception.BadRequestException;
 import roomescape.member.dto.request.AuthRequest.LoginRequest;
-import roomescape.member.dto.request.LoginMember;
 import roomescape.member.entity.Member;
 import roomescape.member.repository.MemberRepository;
